my game as default
